pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "parsley-coco"
version = "0.1.22"
description = "A parser combining dacite and argparse"
authors = [
    { name="Victor Gabillon", email="victorgabillon@gmail.com" }
]
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.12"
dependencies = [
    "PyYAML>=6.0.2",
    "dacite>=1.9.2",
    "click==8.1.8"
]


[project.optional-dependencies]
test = ["pytest>=7.0","types-PyYAML"]


[project.urls]
"Homepage" = "https://github.com/victorgabillon/parsley"


[tool.setuptools]
packages = ["parsley_coco"]

[tool.pytest.ini_options]
testpaths = [
    "parsley_coco",
    "tests_parsley"
]
